Insert a slide

Function writing a quarto revealjs slide according to the user's specifications.

add_slide(
  title = NA,
  section = FALSE,
  color = "#FFFFFF",
  gradient = NA,
  size = "cover",
  opacity = 1,
  trans_data = "fade",
  trans_back = "slide",
  image_url = NA,
  image_position = "center",
  image_repeat = "no-repeat",
  video_url = NA,
  video_loop = "false",
  video_mute = "true",
  iframe_url = NA,
  iframe_interactive = "false",
  classes = NA,
  visible = TRUE
)

Arguments

title

Character. Title of the slide. If NA, the slide does not include separation.

section

Logical. True if this is a section-title slide.

color

Character. Background color.

gradient

Character. Specify a css gradient as background instead of the background color (e.g. "-webkit-radial-gradient(top, circle cover, #246 0%, #013 80%)").

size

Character. Whether the background should be "contain", "cover", or a specific size (e.g. "100px")

opacity

Integer. From 0 for totally transparent to 1 for totally opaque

trans_data

Character. Type of transition for data: fade, slide, convex, concave, or zoom.

trans_back

Character. Type of transition for the background: fade, slide, convex, concave, or zoom.

image_url

Character. URL of the background image or css specification for complex backgrounds.

image_position

Character. Whether the background should be displayed at the "top", "right", "bottom", "left", "center", a=or another position.

image_repeat

Character. Whether the image should "no-repeat", "repeat", "repeat-x" or repeat-y".

video_url

Character. URL of the background video.

video_loop

Character. "true" if the video should loop, "false" otherwise.

video_mute

Character. "true" if the video should be muted, "false" otherwise.

iframe_url

Character. URL of the background website.

iframe_interactive

Character. "true" if the iframe should be interactive, "false" otherwise.

classes

Character vector. Specifiy the names of other formatting classes which should be applied to the slide.

visible

Logical. FALSE if the slide should be hidden.

Value

Character. Write the first rows of a RevealJS slide formatted for quarto.

Author

Nicolas Mangin